fixup! [#50] Trigger a compil. failure on missing @​Self
