+++ jax-rs-linker/src/main/java/fr/vidal/oss/jax_rs_linker/LinkerAnnotationProcessor.java
@@ -1,20 +1,17 @@
 package fr.vidal.oss.jax_rs_linker;
 
 import com.google.auto.service.AutoService;
-import com.google.common.base.Joiner;
-import com.google.common.base.Predicate;
 import com.google.common.collect.FluentIterable;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.LinkedHashMultimap;
 import com.google.common.collect.Multimap;
-import com.google.common.collect.Ordering;
 import fr.vidal.oss.jax_rs_linker.api.Self;
 import fr.vidal.oss.jax_rs_linker.api.SubResource;
 import fr.vidal.oss.jax_rs_linker.functions.OptionalFunctions;
-import fr.vidal.oss.jax_rs_linker.functions.ToMapKey;
 import fr.vidal.oss.jax_rs_linker.model.ClassName;
 import fr.vidal.oss.jax_rs_linker.model.Mapping;
 import fr.vidal.oss.jax_rs_linker.parser.ElementParser;
+import fr.vidal.oss.jax_rs_linker.parser.ResourceGraphValidator;
 import fr.vidal.oss.jax_rs_linker.predicates.OptionalPredicates;
 import fr.vidal.oss.jax_rs_linker.writer.DotFileWriter;
 import fr.vidal.oss.jax_rs_linker.writer.LinkerWriter;
@@ -33,23 +30,19 @@ import javax.lang.model.SourceVersion;
 import javax.lang.model.element.TypeElement;
 import java.io.IOException;
 import java.util.Collection;
-import java.util.Map;
 import java.util.Set;
 
 import static com.google.common.base.Predicates.notNull;
 import static com.google.common.base.Throwables.propagate;
 import static com.google.common.collect.Sets.newHashSet;
-import static fr.vidal.oss.jax_rs_linker.errors.CompilationError.MISSING_SELF;
 import static fr.vidal.oss.jax_rs_linker.functions.JavaxElementToMappings.intoOptionalMapping;
 import static fr.vidal.oss.jax_rs_linker.functions.MappingToClassName.INTO_CLASS_NAME;
 import static fr.vidal.oss.jax_rs_linker.functions.MappingToPathParameters.TO_PATH_PARAMETERS;
 import static fr.vidal.oss.jax_rs_linker.functions.MappingToQueryParameters.TO_QUERY_PARAMETERS;
 import static fr.vidal.oss.jax_rs_linker.functions.TypeElementToElement.intoElement;
 import static fr.vidal.oss.jax_rs_linker.predicates.ElementHasKind.byKind;
-import static fr.vidal.oss.jax_rs_linker.predicates.HasSelfMapping.HAS_SELF;
 import static javax.lang.model.SourceVersion.latest;
 import static javax.lang.model.element.ElementKind.METHOD;
-import static javax.tools.Diagnostic.Kind.ERROR;
 
 
 @AutoService(Processor.class)
@@ -61,8 +54,8 @@ public class LinkerAnnotationProcessor extends AbstractProcessor {
 
     private final Multimap<ClassName, Mapping> elements = LinkedHashMultimap.create();
     private ResourceFileWriters resourceFiles;
-    private Messager messager;
     private ElementParser elementParser;
+    private ResourceGraphValidator validator;
 
     @Override
     public Set<String> getSupportedOptions() {
@@ -85,8 +78,9 @@ public class LinkerAnnotationProcessor extends AbstractProcessor {
     @Override
     public void init(ProcessingEnvironment processingEnv) {
         super.init(processingEnv);
+        Messager messager = processingEnv.getMessager();
         resourceFiles = new ResourceFileWriters(processingEnv.getFiler());
-        messager = processingEnv.getMessager();
+        validator = new ResourceGraphValidator(messager);
         elementParser = new ElementParser(
             messager,
             processingEnv.getTypeUtils()
@@ -107,22 +101,12 @@ public class LinkerAnnotationProcessor extends AbstractProcessor {
                 .filter(notNull())
                 .index(INTO_CLASS_NAME);
 
-        if (validateMappings(roundElements)) {
+        if (validator.validateMappings(roundElements)) {
             tryGenerateSources(roundElements);
             tryExportGraph(roundEnv);
         }
 
         return false;
-    }
-
-    private boolean validateMappings(Multimap<ClassName, Mapping> roundElements) {
-        Collection<ClassName> classesWithoutSelf = classesWithoutSelf(roundElements);
-        if (!classesWithoutSelf.isEmpty()) {
-            String classes = Joiner.on(System.lineSeparator() + "\t - ").join(classesWithoutSelf);
-            messager.printMessage(ERROR, MISSING_SELF.format(classes));
-            return false;
-        }
-        return true;
     }
 
     private void tryGenerateSources(Multimap<ClassName, Mapping> roundElements) {
@@ -182,19 +166,4 @@ public class LinkerAnnotationProcessor extends AbstractProcessor {
         ClassName generatedEnum = className.append("QueryParameters");
         new QueryParamsEnumWriter(processingEnv.getFiler()).write(generatedEnum, mappings);
     }
-
-    private Collection<ClassName> classesWithoutSelf(Multimap<ClassName, Mapping> roundElements) {
-        return FluentIterable.from(roundElements.asMap().entrySet())
-            .filter(new Predicate<Map.Entry<ClassName, Collection<Mapping>>>() {
-                @Override
-                public boolean apply(Map.Entry<ClassName, Collection<Mapping>> classMappings) {
-                    return !FluentIterable.from(classMappings.getValue())
-                        .firstMatch(HAS_SELF)
-                        .isPresent();
-                }
-            })
-            .transform(ToMapKey.<ClassName>intoKey())
-            .toSortedSet(Ordering.<ClassName>natural());
-    }
-
 }
